Controllers
Implemented bookController to simplify bookRoutes.
Also learned about Reavealing Module pattern. (All methods and variables are kept private until they are explicitly exposed; usually through an object literal returned by the closure from which it’s defined.)
